requirements.txt
torch>=2.1.0
transformers==4.54.1
accelerate>=0.30.0
datasets>=2.18.0
pillow
numpy
pandas
pyyaml
tqdm
matplotlib
seaborn
einops
sentencepiece
tiktoken
protobuf
safetensors
huggingface_hub

open_clip_torch>=2.24.0
timm>=0.9.16

tensorboard
wandb
